% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_day_plot}
\alias{all_day_plot}
\title{Histopathology Plot}
\format{
ggplot object
}
\usage{
all_day_plot
}
\description{
Cecal and colonic histopathology score results at the last day of the
experimental timeline.
}
\keyword{datasets}
