% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baiH}
\alias{baiH}
\title{baiH Plot}
\format{
ggplot object
}
\usage{
baiH
}
\description{
Fecal stool sample baiH functional potential (secondary bile acid producer) across
experimental timeline.
}
\keyword{datasets}
