% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baiH_plot}
\alias{baiH_plot}
\title{baiH Statistical Results Plot}
\format{
ggplot object
}
\usage{
baiH_plot
}
\description{
Non-parametric ANOVA statistical results for baiH plot.
}
\keyword{datasets}
