% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baiI}
\alias{baiI}
\title{baiI Plot}
\format{
ggplot object
}
\usage{
baiI
}
\description{
Fecal stool sample baiI functional potential (secondary bile acid producer) across
experimental timeline.
}
\keyword{datasets}
