% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bile_plot}
\alias{bile_plot}
\title{Bile Acid Concentration Plot}
\format{
ggplot object
}
\usage{
bile_plot
}
\description{
Cecal sample primary and secondary bile acid concentrations at the last day of the
experimental timeline, broken up by the name of the bile acid.
}
\keyword{datasets}
