% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buk_plot}
\alias{buk_plot}
\title{Buk Statistical Results Plot}
\format{
ggplot object
}
\usage{
buk_plot
}
\description{
Non-parametric ANOVA statistical results for butyrate kinase plot.
}
\keyword{datasets}
