% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{but_plot}
\alias{but_plot}
\title{But Statistical Results Plot}
\format{
ggplot object
}
\usage{
but_plot
}
\description{
Non-parametric ANOVA statistical results for butyryl-coA transferase plot.
}
\keyword{datasets}
