% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butyrate_kinase}
\alias{butyrate_kinase}
\title{Butyrate Kinase Plot}
\format{
ggplot object
}
\usage{
butyrate_kinase
}
\description{
Fecal stool sample butyrate kinase (buk) functional potential (butyrate producer) across
experimental timeline.
}
\keyword{datasets}
