% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butyrateOnly_plot}
\alias{butyrateOnly_plot}
\title{Cecal Concentration of Butyrate}
\format{
ggplot object
}
\usage{
butyrateOnly_plot
}
\description{
Cecal butyrate concentrations at day 3 for all diets. This is a shortened version of `metab_plot`.
}
\keyword{datasets}
