% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butyryl_coa_transferase}
\alias{butyryl_coa_transferase}
\title{Butyryl CoA Transferase Plot}
\format{
ggplot object
}
\usage{
butyryl_coa_transferase
}
\description{
Fecal stool sample butyryl-coA-transferase (but) functional potential (butyrate producer) across
experimental timeline.
}
\keyword{datasets}
