% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cecal_hypox_plot}
\alias{cecal_hypox_plot}
\title{Cecal Hypoxia and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
cecal_hypox_plot
}
\description{
Cecal sample histopathology scoring components correlated with hypoxia fluorescence levels
at the last day of the experimental timeline.
}
\keyword{datasets}
