% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cecum_histo_plot}
\alias{cecum_histo_plot}
\title{Cecal Histopathology Scores Only Plot}
\format{
ggplot object
}
\usage{
cecum_histo_plot
}
\description{
Cecal histopathology score results at the last day of the experimental timeline.
}
\keyword{datasets}
