% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cecum_plot}
\alias{cecum_plot}
\title{Cecal Histopathology Scoring Components Plot}
\format{
ggplot object
}
\usage{
cecum_plot
}
\description{
Cecal histopathology score results at the last day of the experimental timeline broken up
by the respective scoring components.
}
\keyword{datasets}
