% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_histo_plot}
\alias{colon_histo_plot}
\title{Colon Histopathology Scores Only Plot}
\format{
ggplot object
}
\usage{
colon_histo_plot
}
\description{
Colon histopathology score results at the last day of the experimental timeline.
}
\keyword{datasets}
