% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_hypox_plot}
\alias{colon_hypox_plot}
\title{Colon Hypoxia and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
colon_hypox_plot
}
\description{
Colon sample histopathology scoring components correlated with hypoxia fluorescence levels
at the last day of the experimental timeline.
}
\keyword{datasets}
