% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_plot}
\alias{colon_plot}
\title{Colon Histopathology Scoring Components Plot}
\format{
ggplot object
}
\usage{
colon_plot
}
\description{
Colon histopathology score results at the last day of the experimental timeline broken up
by the respective scoring components.
}
\keyword{datasets}
