% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colony_bloodOnly_plot}
\alias{colony_bloodOnly_plot}
\title{Blood Culture Colony Count Plot}
\format{
ggplot object
}
\usage{
colony_bloodOnly_plot
}
\description{
Colony counts for blood cultures grown aerobically and anaerobically.
}
\keyword{datasets}
