% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plots.R
\name{correlation_plots}
\alias{correlation_plots}
\title{Creates correlation comparison plots}
\usage{
correlation_plots(
  input_table,
  x_value,
  y_value,
  x_transform = FALSE,
  y_transform = FALSE,
  point_alpha = 0.4,
  regression_method = "lm",
  facet_rows,
  facet_cols,
  row_labs,
  col_labs,
  x_name,
  y_name,
  title_content
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{x_value}{The column name of what you want on the x-axis (as a string)}

\item{y_value}{The column name of what you want on the y-axis (as a string)}

\item{x_transform}{Applies log10 transformation to x-axis values, default is FALSE}

\item{y_transform}{Applies log10 transformation to y-axis values, default is FALSE}

\item{point_alpha}{Enter a numeric value between 0 and 1 to change the transparency of your points, default is 0.4}

\item{regression_method}{Defines which line of best fit method you want
(options: 'lm', 'glm', 'gam', 'loess', or 'none'), default is 'lm'}

\item{facet_rows}{The column name that you want to facet your plot ROWS by (as a string)}

\item{facet_cols}{The column name that you want to facet your plot COLUMNS by (as a string)}

\item{row_labs}{A list of new facet ROW labels in the labeller format
ex: new_facet_labs <- c('names',
                        'I want')
    names(new_facet_labs) <- c('old names',
                               'to be replaced')}

\item{col_labs}{A list of new facet COLUMN labels in the labeller format (same as above)}

\item{x_name}{The name of your x-axis (as a string)}

\item{y_name}{The name of your y-axis (as a string)}

\item{title_content}{The title of your plot (as a string)}
}
\description{
`correlation_plots()` returns a plot faceted by two variables, gives you the option to apply a log10
transformation to the x and y-axis, and the line of best fit method is alterable.
}
\examples{
# coming soon!

}
