% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dca_sep_plot}
\alias{dca_sep_plot}
\title{Cecal Bile Acid Concentration}
\format{
ggplot object
}
\usage{
dca_sep_plot
}
\description{
Cecal DCA (secondary bile acid), C. difficile inhibiting bile acid, and C. difficile promoting
bile acid concentrations at day 3. P-value significance is overlaid on the plot.
}
\keyword{datasets}
