% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diet_plot_final}
\alias{diet_plot_final}
\title{Mouse Survival Plot}
\format{
ggplot object
}
\usage{
diet_plot_final
}
\description{
Survival results per diet for the experiment.
}
\keyword{datasets}
