% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dil_plot}
\alias{dil_plot}
\title{Cecal Diluted Toxin Concentration Plot}
\format{
ggplot object
}
\usage{
dil_plot
}
\description{
Cecal diluted toxin concentrations at the last day of the experimental timeline.
}
\keyword{datasets}
