% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_df}
\alias{dist_df}
\title{Principal Coordinates for Exercise Data}
\format{
A dataframe with 100 rows and 3 columns:
\describe{
\item{sampleid}{IDs for each individual sample}
\item{PC1}{Principal coordinates for the first axis}
\item{PC2}{Principal coordinates for the second axis}
}
}
\usage{
dist_df
}
\description{
Generated data for the purpose of package testing and function examples. This dataframe can be combined with
the `exercise_df` as they have the same sample IDs. Note: this data is completely fake and was randomly generated
in R based off of healthy ranges for each value. These samples were not obtained by sampling actual people or by running
an actual UniFrac dissimilarity measurement.
}
\keyword{datasets}
