% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{everything_plot}
\alias{everything_plot}
\title{New Experiment Positive Culture and Survival Results}
\format{
ggplot object
}
\usage{
everything_plot
}
\description{
A summary plot of positive culture results per anatomical location (spleen, liver, or blood) at day 3
combined with whether the associated mouse survived or not. Also includes the microbes found in the
positive cultures.
}
\keyword{datasets}
