% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{everything_stats}
\alias{everything_stats}
\title{New Experiment Positive Culture and Survival Statistical Results}
\format{
ggplot object
}
\usage{
everything_stats
}
\description{
Non-parametric ANOVA pairwise comparisons associated with `everything_plot`. Done between diets and
vendors for positive culture and/or survival.
}
\keyword{datasets}
