% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercise_df}
\alias{exercise_df}
\title{Exercise Data}
\format{
A dataframe with 100 rows and 10 columns:
\describe{
\item{sampleid}{IDs for each individual sample}
\item{sex}{Sex of each individual}
\item{location}{The state the individual lives in}
\item{heart_rate}{The average resting heart rate of each individual}
\item{exercise_mins}{The average amount of time spent exercising per workout (in minutes)}
\item{exercise_weekly}{The average number of times an individual works out per week}
\item{predominant_activity}{The activity most often used in a workout}
\item{bp_systolic}{The average systolic blood pressure measurement of each individual}
\item{bp_diastolic}{The average diastolic blood pressure measurement of each individual}
\item{bmi}{The BMI of each individual}
}
}
\usage{
exercise_df
}
\description{
Generated data for the purpose of package testing and function examples. Note: this data is completely
fake and was randomly generated in R based off of healthy ranges for each value. These samples were not
obtained by sampling actual people.
}
\keyword{datasets}
