% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_once_plots.R
\name{facet_once_plots}
\alias{facet_once_plots}
\title{Create faceted boxplot including statistical results}
\usage{
facet_once_plots(
  input_table,
  stat_on_plot = TRUE,
  x_value,
  y_value,
  grouped_by,
  point_alpha,
  x_labels,
  facet_by,
  facet_labels,
  stat_table = NULL,
  x_name,
  y_name,
  title_content
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{stat_on_plot}{Uses stat_table to overlay p-value significance on plot, default is TRUE}

\item{x_value}{The column name of what you want on the x-axis (as a string)}

\item{y_value}{The column name of what you want on the y-axis (as a string)}

\item{grouped_by}{The column name of how you want your boxplots to be grouped, typically is the same as
the x-axis column name (as a string)}

\item{point_alpha}{Enter a numeric value between 0 and 1 to change the transparency of your points}

\item{x_labels}{A list of new labels for x-axis values (requires a discrete variable)}

\item{facet_by}{The column name that you want to facet your plot by (as a string)}

\item{facet_labels}{A list of new facet labels in the labeller format
ex: new_facet_labs <- c('names',
                        'I want')
    names(new_facet_labs) <- c('old names',
                               'to be replaced')}

\item{stat_table}{The name of your dataframe or tibble containing statistical results (specifically containing columns for:
1. p-value significance - named 'p.adj.signif'
2. y-position - added via add_y_position())}

\item{x_name}{The name of your x-axis (as a string)}

\item{y_name}{The name of your y-axis (as a string)}

\item{title_content}{The title of your plot (as a string)}
}
\description{
`facet_once_plots()` returns a plot faceted by a single variable with data points, boxes, and contains an option
to include p-value significance.
}
\examples{
# example code


}
