% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_twice_plots.R
\name{facet_twice_plots}
\alias{facet_twice_plots}
\title{Creates a twice faceted boxplot}
\usage{
facet_twice_plots(
  input_table,
  x_value,
  y_value,
  y_transform = TRUE,
  x_labels,
  box_group_by,
  line_group_by,
  point_alpha = 0.4,
  facet_rows,
  facet_cols,
  row_labs,
  col_labs,
  x_name,
  y_name,
  title_content
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{x_value}{The column name of what you want on the x-axis (as a string)}

\item{y_value}{The column name of what you want on the y-axis (as a string)}

\item{y_transform}{Applies log10 transformation to y-axis values, default is TRUE}

\item{x_labels}{A list of new labels for x-axis values (requires a continuous variable)}

\item{box_group_by}{The column name of how you want your boxplots to be grouped, typically is the same as
the x-axis column name (as a string)}

\item{line_group_by}{The column name of how you want to be your lines to be grouped (i.e. by sample ID)}

\item{point_alpha}{Enter a numeric value between 0 and 1 to change the transparency of your points, default is 0.4}

\item{facet_rows}{The column name that you want to facet your plot ROWS by (as a string)}

\item{facet_cols}{The column name that you want to facet your plot COLUMNS by (as a string)}

\item{row_labs}{A list of new facet ROW labels in the labeller format
ex: new_facet_labs <- c('names',
                        'I want')
    names(new_facet_labs) <- c('old names',
                               'to be replaced')}

\item{col_labs}{A list of new facet COLUMN labels in the labeller format (same as above)}

\item{x_name}{The name of your x-axis (as a string)}

\item{y_name}{The name of your y-axis (as a string)}

\item{title_content}{The title of your plot (as a string)}
}
\description{
`facet_twice_plots()` is similar to `facet_once_plots()` but gives you the option to facet by two variables
and apply a log10 transformation to the y-axis. Best for use with longitudinal data.
}
\section{TEST}{

}

\examples{
# example code

}
