% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faith_combStats_plot}
\alias{faith_combStats_plot}
\title{Cecal Microbiome Faith's PD Correlated to DCA and Butyrate Concentrations}
\format{
ggplot object
}
\usage{
faith_combStats_plot
}
\description{
Cecal microbiome Faith's PD at day 3 correlated with DCA (secondary bile acid) and butyrate concentrations
at the same time point. Significant p-values and r2 values are overlaid on the plot. Points are
colored by diet rather than being faceted.
}
\keyword{datasets}
