% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faith_plot}
\alias{faith_plot}
\title{Faith's PD Plot (Alpha Diversity)}
\format{
ggplot object
}
\usage{
faith_plot
}
\description{
Fecal stool sample faith's PD scores across experimental timeline.
}
\keyword{datasets}
