% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faith_stat_vis}
\alias{faith_stat_vis}
\title{Faith's PD Statistical Results Plot (Alpha Diversity)}
\format{
ggplot object
}
\usage{
faith_stat_vis
}
\description{
Non-parametric ANOVA for fecal stool sample faith's PD scores across experimental timeline.
}
\keyword{datasets}
