% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family_abun1}
\alias{family_abun1}
\title{Microbe Family Relative Abundance Plot}
\format{
ggplot object
}
\usage{
family_abun1
}
\description{
Fecal stool sample microbe relative abundances across experimental timeline, sectioned out by Family.
}
\keyword{datasets}
