% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genus_plot1}
\alias{genus_plot1}
\title{Microbe Genera Relative Abundance Plot}
\format{
ggplot object
}
\usage{
genus_plot1
}
\description{
Fecal stool sample microbe relative abundances across experimental timeline, sectioned out by Genus.
}
\keyword{datasets}
