% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genusAbun_plots_together_withLabs}
\alias{genusAbun_plots_together_withLabs}
\title{Microbe Genus Relative Abundance Plot}
\format{
ggplot object
}
\usage{
genusAbun_plots_together_withLabs
}
\description{
Fecal stool sample microbe relative abundances across experimental timeline, sectioned out by Genus.
Includes obligate anaerobes and potential pathogens. Faceted by diet.
}
\keyword{datasets}
