% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_acetic_plot}
\alias{histo_acetic_plot}
\title{Acetate Concentration and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_acetic_plot
}
\description{
Cecal concentrations of acetate (acetic acid) correlated with cecal and colonic histopathology scores at the
last day of the experimental timeline.
}
\keyword{datasets}
