% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_butanoic_plot}
\alias{histo_butanoic_plot}
\title{Butyrate Concentration and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_butanoic_plot
}
\description{
Cecal concentrations of butyrate (n-butanoic acid) correlated with cecal and colonic histopathology scores at the
last day of the experimental timeline.
}
\keyword{datasets}
