% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_inhibitors_plot}
\alias{histo_inhibitors_plot}
\title{Bile Acid Inhibitors and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_inhibitors_plot
}
\description{
Cecal bile acid inhibitor concentrations correlated with cecal and colonic histopathology scores
at the last day of the experimental timeline.
}
\keyword{datasets}
