% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_metabCecum_plot}
\alias{histo_metabCecum_plot}
\title{Metabolite Concentration and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_metabCecum_plot
}
\description{
Cecal metabolite concentrations (acetic acid, n-butanoic acid, propanoic acid) correlated with
cecal histopathology scores at the last day of the experimental timeline.
}
\keyword{datasets}
