% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_perc_plot}
\alias{histo_perc_plot}
\title{Cecal Histopathology Scores Component Percentages Plot}
\format{
ggplot object
}
\usage{
histo_perc_plot
}
\description{
Cecal histopathology score component percentage results at the last day of the experimental timeline.
This is based on the percentage of "bad" histopathology scores received per diet.
}
\keyword{datasets}
