% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_promoters_plot}
\alias{histo_promoters_plot}
\title{Bile Acid Promoters and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_promoters_plot
}
\description{
Cecal bile acid promoter concentrations correlated with cecal and colonic histopathology scores
at the last day of the experimental timeline.
}
\keyword{datasets}
