% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_propanoic_plot}
\alias{histo_propanoic_plot}
\title{Propanoate Concentration and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_propanoic_plot
}
\description{
Cecal concentrations of propanoate (propanoic acid) correlated with cecal and colonic histopathology scores at the
last day of the experimental timeline.
}
\keyword{datasets}
