% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_ratio_plot}
\alias{histo_ratio_plot}
\title{Bile Acid Ratio and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
histo_ratio_plot
}
\description{
Cecal bile acid promoter:inhibitor ratios correlated with cecal and colonic histopathology scores
at the last day of the experimental timeline.
}
\keyword{datasets}
