% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histo_ratio_stat_plot}
\alias{histo_ratio_stat_plot}
\title{Bile Acid Ratio and Histopathology Score Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
histo_ratio_stat_plot
}
\description{
Cecal bile acid promoter:inhibitor ratios correlated with cecal and colonic histopathology scores
statistical results. Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
