% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_cecum_plot}
\alias{hypox_cecum_plot}
\title{Hypoxia and Histopathology Score Correlation Plot (in the cecum ONLY)}
\format{
ggplot object
}
\usage{
hypox_cecum_plot
}
\description{
Cecal histopathology scores correlated with cecal hypoxia fluorescence levels
at the last day of the experimental timeline.
}
\keyword{datasets}
