% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_cecumLocation_plot}
\alias{hypox_cecumLocation_plot}
\title{Cecal Hypoxia Plot (diet on x-axis, faceted by GI location)}
\format{
ggplot object
}
\usage{
hypox_cecumLocation_plot
}
\description{
Cecal hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
