% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_colonStats_plot}
\alias{hypox_colonStats_plot}
\title{Hypoxia and Histopathology Score Correlation Plot (in the colon ONLY)}
\format{
ggplot object
}
\usage{
hypox_colonStats_plot
}
\description{
Colonic histopathology scores correlated with colonic hypoxia fluorescence levels
at the last day of the experimental timeline. P-value significance is overlaid on the plot.
}
\keyword{datasets}
