% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_location_plot}
\alias{hypox_location_plot}
\title{Hypoxia Plot (diet on x-axis, faceted by GI location)}
\format{
ggplot object
}
\usage{
hypox_location_plot
}
\description{
Hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
