% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_tcda_plot}
\alias{hypox_tcda_plot}
\title{Hypoxia and TcdA Toxin Correlation Plot}
\format{
ggplot object
}
\usage{
hypox_tcda_plot
}
\description{
Cecal hypoxia fluorescence levels correlated with cecal TcdA toxin concentration
at the last day of the experimental timeline.
}
\keyword{datasets}
