% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_tcdb_plot}
\alias{hypox_tcdb_plot}
\title{Hypoxia and TcdB Toxin Correlation Plot}
\format{
ggplot object
}
\usage{
hypox_tcdb_plot
}
\description{
Cecal hypoxia fluorescence levels correlated with cecal TcdB toxin concentration
at the last day of the experimental timeline.
}
\keyword{datasets}
