% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypox_toxin_stat_plot}
\alias{hypox_toxin_stat_plot}
\title{Hypoxia and Toxin Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
hypox_toxin_stat_plot
}
\description{
Cecal hypoxia fluorescence levels correlated with cecal toxin concentration
statistical results. Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
