% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypoxia_cecum_plot}
\alias{hypoxia_cecum_plot}
\title{Cecalk Hypoxia Levels}
\format{
ggplot object
}
\usage{
hypoxia_cecum_plot
}
\description{
Cecal hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
