% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypoxia_dietLocation_plot}
\alias{hypoxia_dietLocation_plot}
\title{Hypoxia Plot (GI location on x-axis, faceted by diet)}
\format{
ggplot object
}
\usage{
hypoxia_dietLocation_plot
}
\description{
Hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
