% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypoxia_locationDiet_plot}
\alias{hypoxia_locationDiet_plot}
\title{Hypoxia Plot (diet on x-axis, faceted by GI location)}
\format{
ggplot object
}
\usage{
hypoxia_locationDiet_plot
}
\description{
Hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
