% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lachno_abun_plot}
\alias{lachno_abun_plot}
\title{Lachnospiraceae Relative Abundance Plot}
\format{
ggplot object
}
\usage{
lachno_abun_plot
}
\description{
Features the top five most abundant genera in Lachnospiraceae and follows their relative
abundance throughout the experimental timeline, faceted by diet. Is a plot in `genusAbun_plots_together_withLabs`.
}
\keyword{datasets}
