% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Run a Linear Model}
\usage{
linear_model(
  input_table,
  grouped_by,
  multiple_groups = FALSE,
  adjust_method,
  filter_adj_p_value = FALSE,
  formula_left,
  formula_right
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{grouped_by}{The column name you want your data grouped by prior to statistical analysis (as a string or a list of strings).
Typically it's whatever you've faceted your plot by (if you want your stats on your plot).}

\item{multiple_groups}{If TRUE, will know that you've grouped your data by two columns. If FALSE, will know that you've only grouped
your data by one column. Default is FALSE.}

\item{adjust_method}{P-value adjustment method (as a string). Options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", and "none".}

\item{filter_adj_p_value}{If TRUE, will filter your Kruskal Test adjusted p-values <= 0.05 significance and will
filter all non-significant results out of the input table prior to the Dunn's Post Hoc test. Default is FALSE.}

\item{formula_left}{The column name of the variable you want tested on the left side of the formula (as a string).
ex: formula_left ~ formula_right}

\item{formula_right}{The column name of the variable you want tested on the right side of the formula (as a string).
ex: formula_left ~ formula_right}
}
\description{
`linear_model()` returns a dataframe with results of your linear regression analysis.
}
\examples{
linear_model <- linear_model(input_table = exercise_df,
                             grouped_by = c('sex',
                                            'location'),
                             multiple_groups = TRUE,
                             adjust_method = 'BH',
                             filter_adj_p_value = FALSE,
                             formula_left = 'heart_rate',
                             formula_right = 'bmi + predominant_activity')

}
