% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_fib_plot}
\alias{location_fib_plot}
\title{Hypoxia Plot (diet fiber level on x-axis, faceted by GI location)}
\format{
ggplot object
}
\usage{
location_fib_plot
}
\description{
Hypoxia fluorescence levels at the last day of the experimental timeline.
}
\keyword{datasets}
