% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab_bile_stat_plot}
\alias{metab_bile_stat_plot}
\title{Metabolite and Bile Acid Concentration Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
metab_bile_stat_plot
}
\description{
Cecal metabolite concentrations (acetic acid, n-butanoic acid, and propanoic acid) correlated
with cecal bile acid concentrations statistical results. Indicates p-value significance and the
correlation coefficient.
}
\keyword{datasets}
