% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab_inhibitors_plot}
\alias{metab_inhibitors_plot}
\title{Metabolite and Bile Acid Inhibitors Correlation Plot}
\format{
ggplot object
}
\usage{
metab_inhibitors_plot
}
\description{
Cecal metabolite concentrations (acetic acid, n-butanoic acid, and propanoic acid) correlated
with cecal bile acid inhibitors at the last day of the experimental timeline.
}
\keyword{datasets}
