% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab_plot}
\alias{metab_plot}
\title{Metabolite Concentration Plot}
\format{
ggplot object
}
\usage{
metab_plot
}
\description{
Cecal metabolite concentrations (acetic acid, n-butanoic acid, and propanoic acid)
 at the last day of the experimental timeline.
}
\keyword{datasets}
