% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab_stat_vis}
\alias{metab_stat_vis}
\title{Metabolite and Histopathology Score Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
metab_stat_vis
}
\description{
Cecal metabolite concentrations (acetic acid, n-butanoic acid, and propanoic acid) correlated
with histopathology score statistical results.
Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
