% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbeProd_relAbunStats_plot}
\alias{microbeProd_relAbunStats_plot}
\title{Cecal Microbiome Relative Abundances Correlated to DCA and Butyrate Concentrations}
\format{
ggplot object
}
\usage{
microbeProd_relAbunStats_plot
}
\description{
Cecal microbiome relative abundances of key families and genera at day 3 compared to concentrations
of DCA and butyrate at the same day. Significant p-values and r2 values are overlaid on the plot.
Points are colored by diet rather than being faceted.
}
\keyword{datasets}
