% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_d3_genus_relAbun_stat_plot}
\alias{mini_d3_genus_relAbun_stat_plot}
\title{All Experiment Comparison Relative Abundance Stats Plot (between experiment and vendor)}
\format{
ggplot object
}
\usage{
mini_d3_genus_relAbun_stat_plot
}
\description{
Pairwise comparison statistics associated with the relative abundance of potential pathogenic genera
at days -15 and 3 between sets of experiments done (Old Anschutz (2020), U of Arizona, New Anschutz (2024)).
Done via non-parametric ANOVA.
}
\keyword{datasets}
