% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_neat_cecumHisto_stats_plot}
\alias{mini_neat_cecumHisto_stats_plot}
\title{Cecal Toxin and Histopathology Score Correlation Plot (mini version)}
\format{
ggplot object
}
\usage{
mini_neat_cecumHisto_stats_plot
}
\description{
Cecal toxin concentrations correlated with cecal histopathology scores at the last day of
the experimental timeline. P-value significance is overlaid on the plot and only significant
results are pictured.
}
\keyword{datasets}
