% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniDiet_genusAbun_plot}
\alias{miniDiet_genusAbun_plot}
\title{All Experiment Comparison Relative Abundance Plot (colored by diet)}
\format{
ggplot object
}
\usage{
miniDiet_genusAbun_plot
}
\description{
Comparison of the relative abundance of potential pathogenic genera at days -15 and 3 between sets
of experiments done (Old Anschutz (2020), U of Arizona, New Anschutz (2024)). Points are colored by
diet for comparison between experiments.
}
\keyword{datasets}
