% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniVendor_genusAbun_plot}
\alias{miniVendor_genusAbun_plot}
\title{All Experiment Comparison Relative Abundance Plot (colored by vendor)}
\format{
ggplot object
}
\usage{
miniVendor_genusAbun_plot
}
\description{
Comparison of the relative abundance of potential pathogenic genera at days -15 and 3 between sets
of experiments done (Old Anschutz (2020), U of Arizona, New Anschutz (2024)). Points are colored by
mouse vendor since it differed based on experiment.
}
\keyword{datasets}
