% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neat_cecumHisto_stats_plot}
\alias{neat_cecumHisto_stats_plot}
\title{Cecal Toxin and Histopathology Score Correlation Plot (full version)}
\format{
ggplot object
}
\usage{
neat_cecumHisto_stats_plot
}
\description{
Cecal toxin concentrations correlated with cecal histopathology scores at the last day of
the experimental timeline. P-value significance is overlaid on the plot and all
results are pictured, whether they're significant or not.
}
\keyword{datasets}
