% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neat_histo_colon_plot}
\alias{neat_histo_colon_plot}
\title{Colon Toxin and Histopathology Score Correlation Plot}
\format{
ggplot object
}
\usage{
neat_histo_colon_plot
}
\description{
Colonic toxin concentrations correlated with colonic histopathology scores at the last day of
the experimental timeline.
}
\keyword{datasets}
