% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neat_plot}
\alias{neat_plot}
\title{Cecal Neat Toxin Concentration Plot}
\format{
ggplot object
}
\usage{
neat_plot
}
\description{
Cecal neat toxin concentrations at the last day of the experimental timeline.
}
\keyword{datasets}
