% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neat_tox_stat_vis}
\alias{neat_tox_stat_vis}
\title{Toxin and Histopathology Score Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
neat_tox_stat_vis
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with histopathology score statistical results.
Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
