% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newExp_avWeight_plot}
\alias{newExp_avWeight_plot}
\title{New Experiment Average Percent Change in Mouse Weight Plot}
\format{
ggplot object
}
\usage{
newExp_avWeight_plot
}
\description{
Average percent change in mouse weight by diet and vendor in the days post infection with C. difficile.
Associated with New Anschutz experiments.
}
\keyword{datasets}
