% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newExp_dietVendor_surv_plot}
\alias{newExp_dietVendor_surv_plot}
\title{New Experiment Survival Plot}
\format{
ggplot object
}
\usage{
newExp_dietVendor_surv_plot
}
\description{
Mouse survival curves by diet and vendor in the days post infection with C. difficile.
Associated with New Anschutz experiments.
}
\keyword{datasets}
