% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noDiet_histoCecum_toxStats_plot}
\alias{noDiet_histoCecum_toxStats_plot}
\title{Toxin and Histopathology Score Correlation Plot (not faceted by diet)}
\format{
ggplot object
}
\usage{
noDiet_histoCecum_toxStats_plot
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with histopathology score. Significant p-values
and r2 values are overlaid on the plot. Points are colored by diet rather than being faceted.
}
\keyword{datasets}
