% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notButy_plot}
\alias{notButy_plot}
\title{Cecal Concentration of Acetate and Propionate}
\format{
ggplot object
}
\usage{
notButy_plot
}
\description{
Cecal acetate and propionate concentrations at day 3 for all diets. This is a shortened version
of `metab_plot`, its final third is `butyrateOnly_plot`.
}
\keyword{datasets}
