% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_plot}
\alias{overall_plot}
\title{Bile Acid Concentration Plot (promoters and inhibitors ONLY)}
\format{
ggplot object
}
\usage{
overall_plot
}
\description{
Cecal sample primary and secondary bile acid concentrations at the last day of the
experimental timeline, broken up by whether they promote or inhibit C. difficile growth.
}
\keyword{datasets}
