% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panelE_fig3_redoWithlab}
\alias{panelE_fig3_redoWithlab}
\title{Cecal Butyrate Concentration Correlated to Toxin and Histopathology Results}
\format{
ggplot object
}
\usage{
panelE_fig3_redoWithlab
}
\description{
Cecal butyrate concentration at day 3 compared to C. difficile toxins (TcdA/TcdB) and cecal histopathology
scores at the same time point. Significant p-values and r2 values are overlaid on the plot.
Points are colored by diet rather than being faceted.
}
\keyword{datasets}
