% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_abun_plot}
\alias{path_abun_plot}
\title{Potential Pathogen Relative Abundance Plot}
\format{
ggplot object
}
\usage{
path_abun_plot
}
\description{
Features five potential pathogen genera and follows their relative abundance throughout the experimental
timeline, faceted by diet. Is a plot in `genusAbun_plots_together_withLabs`.
}
\keyword{datasets}
