% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa_plots.R
\name{pcoa_plots}
\alias{pcoa_plots}
\title{Create PCoA plot}
\usage{
pcoa_plots(
  input_table,
  x_value,
  y_value,
  facet_once = FALSE,
  facet_twice = FALSE,
  point_fill,
  point_alpha = 0.7,
  point_size = 2,
  distiller_palette = "Spectral",
  facet_rows = NULL,
  facet_cols = NULL,
  row_labs = NULL,
  col_labs = NULL,
  legend_name,
  x_name,
  y_name,
  title_content
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{x_value}{The column name of what you want on the x-axis (as a string)}

\item{y_value}{The column name of what you want on the y-axis (as a string)}

\item{facet_once}{If TRUE, will allow you to facet by one variable. Put variable column name (as a string) into the
`facet_cols` parameter, default is FALSE}

\item{facet_twice}{If TRUE, will allow you to facet by two variable. Put variable column names (as a string) into the
`facet_cols` and `facet_rows` parameters, default is FALSE}

\item{point_fill}{The column name of what you want your points to be colored/filled by (as a string)}

\item{point_alpha}{Enter a numeric value between 0 and 1 to change the transparency of your points, default is 0.7}

\item{point_size}{Enter a numeric value to change the size of your points, default is 2}

\item{distiller_palette}{The name of the `scale_fill_distiller()` palette you want your points colored by (as a string),
default is 'Spectral'}

\item{facet_rows}{The column name that you want to facet your plot ROWS by (as a string)}

\item{facet_cols}{The column name that you want to facet your plot COLUMNS by (as a string)}

\item{row_labs}{A list of new facet ROW labels in the labeller format
ex: new_facet_labs <- c('names',
                        'I want')
    names(new_facet_labs) <- c('old names',
                               'to be replaced')}

\item{col_labs}{A list of new facet COLUMN labels in the labeller format (same as above)}

\item{legend_name}{The name of your legend (as a string)}

\item{x_name}{The name of your x-axis (as a string)}

\item{y_name}{The name of your y-axis (as a string)}

\item{title_content}{The title of your plot (as a string)}
}
\description{
`pcoa_plots()` returns a PCoA plot based on your unweighted/weighted UniFrac results. Can be faceted once or twice
by desired metadata variables.
}
\examples{
# example code


}
