% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poster_abun1}
\alias{poster_abun1}
\title{Microbe Family Relative Abundance Plot (mini version)}
\format{
ggplot object
}
\usage{
poster_abun1
}
\description{
Fecal stool sample microbe relative abundances across experimental timeline, sectioned out by Family.
Shortened to include only 5 families, works well for usage in presentations and/or posters.
}
\keyword{datasets}
