% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poster_abun2}
\alias{poster_abun2}
\title{Blood Culture Genera Relative Abundance Plot}
\format{
ggplot object
}
\usage{
poster_abun2
}
\description{
Blood culture sample microbe relative abundances, sectioned out by Genus.
}
\keyword{datasets}
