% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poster_stat_plot1}
\alias{poster_stat_plot1}
\title{Microbe Family Relative Abundance Statistical Results Plot (mini version)}
\format{
ggplot object
}
\usage{
poster_stat_plot1
}
\description{
Non-parametric ANOVA results of microbe family relative abundances between diets.
}
\keyword{datasets}
