% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratio_bile_plot}
\alias{ratio_bile_plot}
\title{Bile Acid Ratio Plot}
\format{
ggplot object
}
\usage{
ratio_bile_plot
}
\description{
Cecal sample bile acid promoter:inhibitor ratios at the last day of the
experimental timeline.
}
\keyword{datasets}
