% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rumino_abun_plot}
\alias{rumino_abun_plot}
\title{Ruminococcaeae Relative Abundance Plot}
\format{
ggplot object
}
\usage{
rumino_abun_plot
}
\description{
Features the top four most abundant genera in Ruminococcaceae and follows their relative
abundance throughout the experimental timeline, faceted by diet. Is a plot in `genusAbun_plots_together_withLabs`.
}
\keyword{datasets}
