% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_abun1}
\alias{short_abun1}
\title{Microbe Family Relative Abundance Plot (manuscript version)}
\format{
ggplot object
}
\usage{
short_abun1
}
\description{
Fecal stool sample microbe relative abundances across experimental timeline, sectioned out by Family.
This version is used in a figure in the paper manuscript.
}
\keyword{datasets}
