% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heat_plot.R
\name{stat_heat_plot}
\alias{stat_heat_plot}
\title{Creates a statistical visualization}
\usage{
stat_heat_plot(
  input_table,
  x_value,
  y_value,
  x_value_type = NULL,
  y_value_type = NULL,
  pairwise = FALSE,
  correlation = FALSE,
  tile_fill,
  tile_outline_color = "black",
  tile_alpha = 0.6,
  text_value,
  low_color = "blue",
  high_color = "green",
  legend_name,
  x_labels = NULL,
  y_labels = NULL,
  facet_by,
  facet_labels,
  x_name,
  y_name,
  title_content
)
}
\arguments{
\item{input_table}{Expects a dataframe or tibble.}

\item{x_value}{The column name of what you want on the x-axis (as a string)}

\item{y_value}{The column name of what you want on the y-axis (as a string)}

\item{x_value_type}{Is your x-axis value continuous or discrete? Accepts inputs of 'NULL' (you don't want you x-axis re-labeled),
'continuous' or 'discrete'.}

\item{y_value_type}{Is your y-axis value continuous or discrete? Accepts inputs of 'NULL' (you don't want you y-axis re-labeled),
'continuous' or 'discrete'.}

\item{pairwise}{Formats plot for pairwise test results (i.e have group comparisons), default is FALSE}

\item{correlation}{Formats plot for linear regression test results (i.e correlation comparisons), default is FALSE}

\item{tile_fill}{The column name of which numeric value you want fill the tiles (as as string)}

\item{tile_outline_color}{The color you want to outline the individual tiles with, default is BLACK}

\item{tile_alpha}{Enter a numeric value between 0 and 1 to change the transparency of your points, default is 0.6}

\item{text_value}{The column name of the text you want overlaid on your plot tiles (as a string)}

\item{low_color}{Tile numeric low point color, default is BLUE}

\item{high_color}{Tile numeric high point color, default is GREEN}

\item{legend_name}{The name of your legend (as a string)}

\item{x_labels}{A list of new labels for x-axis values (requires a discrete variable)}

\item{y_labels}{A list of new labels for y-axis values (requires a discrete variable)}

\item{facet_by}{The column name that you want to facet your plot by (as a string)}

\item{facet_labels}{A list of new facet labels in the labeller format
ex: new_facet_labs <- c('names',
                        'I want')
    names(new_facet_labs) <- c('old names',
                               'to be replaced')}

\item{x_name}{The name of your x-axis (as a string)}

\item{y_name}{The name of your y-axis (as a string)}

\item{title_content}{The title of your plot (as a string)}
}
\description{
`stat_heat_plot()` returns a heat map based on the statistical results entered.
Most compatible with pairwise or linear regression results.
}
\examples{
# example code

}
