% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stat_plot1}
\alias{stat_plot1}
\title{Microbe Family Relative Abundance Statistical Results Plot (manuscript version)}
\format{
ggplot object
}
\usage{
stat_plot1
}
\description{
Non-parametric ANOVA results of microbe family relative abundances between diets.
This version is used in a figure in the paper manuscript.
}
\keyword{datasets}
