% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surv_stat_vis}
\alias{surv_stat_vis}
\title{Survival Statistical Results Plot}
\format{
ggplot object
}
\usage{
surv_stat_vis
}
\description{
Non-parametric ANOVA results of mouse survival between diets.
}
\keyword{datasets}
