% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tox_acetic}
\alias{tox_acetic}
\title{Acetate and Toxin Concentration Correlation Plot}
\format{
ggplot object
}
\usage{
tox_acetic
}
\description{
Cecal concentrations of acetate (acetic acid) correlated with toxin (TcdA and TcdB) concentrations at the
last day of the experimental timeline.
}
\keyword{datasets}
