% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tox_bile_stat_plot}
\alias{tox_bile_stat_plot}
\title{Toxin and Bile Acid Concentration Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
tox_bile_stat_plot
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with cecal bile acid concentrations statistical
results. Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
