% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tox_butanoic}
\alias{tox_butanoic}
\title{Butyrate and Toxin Concentration Correlation Plot}
\format{
ggplot object
}
\usage{
tox_butanoic
}
\description{
Cecal concentrations of butyrate (n-butanoic acid) correlated with toxin (TcdA and TcdB) concentrations at the
last day of the experimental timeline.
}
\keyword{datasets}
