% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tox_metab_stat_vis}
\alias{tox_metab_stat_vis}
\title{Toxin and Metabolite Concentration Correlation Statistical Results Plot}
\format{
ggplot object
}
\usage{
tox_metab_stat_vis
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with cecal metabolite concentrations (acetic acid, n-butanoic acid, and propanoic acid)
statistical results. Indicates p-value significance and the correlation coefficient.
}
\keyword{datasets}
