% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tox_propanoic}
\alias{tox_propanoic}
\title{Propanoate and Toxin Concentration Correlation Plot}
\format{
ggplot object
}
\usage{
tox_propanoic
}
\description{
Cecal concentrations of propanoate (propanoic acid) correlated with toxin (TcdA and TcdB) concentrations at the
last day of the experimental timeline.
}
\keyword{datasets}
