% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toxin_inhibitors_stats_plot}
\alias{toxin_inhibitors_stats_plot}
\title{Toxin and Bile Acid Inhibitors Correlation Plot}
\format{
ggplot object
}
\usage{
toxin_inhibitors_stats_plot
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with cecal bile acid inhibitors at the last day of the experimental timeline.
Significant p-values are overlaid on the plot.
}
\keyword{datasets}
