% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toxin_promoters_plot}
\alias{toxin_promoters_plot}
\title{Toxin and Bile Acid Promoters Correlation Plot}
\format{
ggplot object
}
\usage{
toxin_promoters_plot
}
\description{
Cecal toxin concentrations (TcdA and TcdB) correlated with cecal bile acid promoters at the last day of the experimental timeline.
}
\keyword{datasets}
