% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unweighted_pcoa}
\alias{unweighted_pcoa}
\title{Unweighted UniFrac PCoA Plot}
\format{
ggplot object
}
\usage{
unweighted_pcoa
}
\description{
Fecal stool sample distances by diet throughout the experimental timeline.
}
\keyword{datasets}
