% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uu_homog_plot}
\alias{uu_homog_plot}
\title{Unweighted UniFrac Microbiome Homogeneity Plot}
\format{
ggplot object
}
\usage{
uu_homog_plot
}
\description{
Fecal stool sample pairwise distances by diet within the experimental time points.
}
\keyword{datasets}
