% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w_homog_stat_vis}
\alias{w_homog_stat_vis}
\title{Weighted UniFrac Microbiome Homogeneity Statistical Results Plot}
\format{
ggplot object
}
\usage{
w_homog_stat_vis
}
\description{
Non-parametric ANOVA of fecal stool sample pairwise distances by diet within the
experimental time points.
}
\keyword{datasets}
