% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w_resil_stat_vis}
\alias{w_resil_stat_vis}
\title{Weighted UniFrac Microbiome Resiliency Statistical Results Plot}
\format{
ggplot object
}
\usage{
w_resil_stat_vis
}
\description{
Non-parametric ANOVA of fecal stool sample pairwise distances by diet between baseline microbiome
and subsequent time points.
}
\keyword{datasets}
