% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weighted_pcoa}
\alias{weighted_pcoa}
\title{Weighted UniFrac PCoA Plot}
\format{
ggplot object
}
\usage{
weighted_pcoa
}
\description{
Fecal stool sample distances by diet throughout the experimental timeline.
}
\keyword{datasets}
