% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wu_homog_plot}
\alias{wu_homog_plot}
\title{Weighted UniFrac Microbiome Homogeneity Plot}
\format{
ggplot object
}
\usage{
wu_homog_plot
}
\description{
Fecal stool sample pairwise distances by diet within the experimental time points.
}
\keyword{datasets}
