% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wu_resil_plot}
\alias{wu_resil_plot}
\title{Weighted UniFrac Microbiome Resiliency Plot}
\format{
ggplot object
}
\usage{
wu_resil_plot
}
\description{
Fecal stool sample pairwise distances by diet between baseline microbiome and subsequent time points.
}
\keyword{datasets}
